+++ tests/unit/station/conftest.py
@@ -108,6 +108,7 @@ def mock_tile_builder_fixture(
         "StartBeamformer",
         "ConfigureIntegratedChannelData",
         "StartAcquisition",
+        "SetCspDownload",
         "StopDataTransmission",
         "StopIntegratedData",
         "StopBeamformer",