+++ tests/unit/station/test_station_device.py
@@ -151,6 +151,38 @@ def station_device_fixture(
     yield test_context.get_sps_station_device()
 
 
+@pytest.fixture(name="on_station_device")
+def on_station_device_fixture(
+    test_context: SpsTangoTestHarnessContext,
+    change_event_callbacks: MockTangoEventCallbackGroup,
+) -> DeviceProxy:
+    """
+    Fixture that returns the SPS station Tango device under test.
+
+    Makes sure the device is ONLINE and ON.
+
+    :param test_context: a Tango test context
+        containing an SPS station and mock subservient devices.
+    :param change_event_callbacks: dictionary of Tango change event
+        callbacks with asynchrony support.
+
+    :returns: the station Tango device under test.
+    """
+    sps_station = test_context.get_sps_station_device()
+    sps_station.subscribe_event(
+        "state",
+        EventType.CHANGE_EVENT,
+        change_event_callbacks["state"],
+    )
+    change_event_callbacks["state"].assert_change_event(DevState.DISABLE)
+    if sps_station.adminMode != AdminMode.ONLINE:
+        sps_station.adminMode = AdminMode.ONLINE
+        change_event_callbacks["state"].assert_change_event(DevState.UNKNOWN)
+        change_event_callbacks["state"].assert_change_event(DevState.ON)
+
+    return sps_station
+
+
 @pytest.fixture(name="daq_device")
 def daq_device_fixture(
     test_context: SpsTangoTestHarnessContext,
@@ -387,59 +419,7 @@ def test_On(
     )
     for i, tile in enumerate(mock_tile_device_proxies):
         last_tile = i == num_tiles - 1
-        if last_tile:
-            num_csp_ingest_configures = 6
-        else:
-            num_csp_ingest_configures = 0
-        num_configures = 6 + num_csp_ingest_configures
-        assert len(tile.Configure40GCore.mock_calls) == num_configures
-        for core in range(2):
-            assert json.loads(
-                tile.Configure40GCore.mock_calls[
-                    num_csp_ingest_configures + (3 * core)
-                ].args[0]
-            ) == {
-                "core_id": core,
-                "arp_table_entry": 0,
-                "source_port": 61648,
-                "destination_ip": (
-                    f"10.0.0.{str(154 + (2 * i) + core)}"
-                    if i != num_tiles - 1
-                    else csp_ingest_address
-                ),
-                "destination_port": csp_ingest_port,
-                "rx_port_filter": csp_ingest_port,
-                "netmask": int(ipaddress.ip_interface(sdn_first_interface).netmask),
-                "gateway_ip": int(ipaddress.ip_address(sdn_gateway)),
-            }
-            assert json.loads(
-                tile.Configure40GCore.mock_calls[
-                    num_csp_ingest_configures + 1 + (3 * core)
-                ].args[0]
-            ) == {
-                "core_id": core,
-                "arp_table_entry": 2,
-                "source_port": 61648,
-                "destination_ip": (
-                    f"10.0.0.{str(154 + (2 * i) + core)}"
-                    if i != num_tiles - 1
-                    else csp_ingest_address
-                ),
-                "destination_port": (
-                    csp_ingest_port + 2 if not last_tile else csp_ingest_port
-                ),
-                "netmask": int(ipaddress.ip_interface(sdn_first_interface).netmask),
-                "gateway_ip": int(ipaddress.ip_address(sdn_gateway)),
-            }
-            assert json.loads(
-                tile.Configure40GCore.mock_calls[
-                    num_csp_ingest_configures + 2 + (3 * core)
-                ].args[0]
-            ) == {
-                "core_id": core,
-                "arp_table_entry": 1,
-                "rx_port_filter": csp_ingest_port + 2,
-            }
+        tile.Configure40GCore.assert_not_called()
         assert len(tile.ConfigureStationBeamformer.mock_calls) == 1
         assert json.loads(tile.ConfigureStationBeamformer.mock_calls[0].args[0]) == {
             "is_first": (i == 0),
@@ -453,10 +433,10 @@ def test_On(
                     "destination_ip": "10.244.170.166",
                     "destination_port": 4660,
                     "source_port": 61648,
-                    "netmask_40g": int(
+                    "netmask_40g": str(
                         ipaddress.ip_interface(sdn_first_interface).netmask
                     ),
-                    "gateway_40g": int(ipaddress.ip_address(sdn_gateway)),
+                    "gateway_40g": sdn_gateway,
                 }
             )
         )
@@ -469,10 +449,10 @@ def test_On(
                     "destination_ip": "10.244.170.166",
                     "source_port": 61648,
                     "destination_port": 4660,
-                    "netmask_40g": int(
+                    "netmask_40g": str(
                         ipaddress.ip_interface(sdn_first_interface).netmask
                     ),
-                    "gateway_40g": int(ipaddress.ip_address(sdn_gateway)),
+                    "gateway_40g": sdn_gateway,
                 }
             )
         )
@@ -685,59 +665,7 @@ def test_Initialise(
 
     for i, tile in enumerate(mock_tile_device_proxies):
         last_tile = i == num_tiles - 1
-        if last_tile:
-            num_csp_ingest_configures = 6
-        else:
-            num_csp_ingest_configures = 0
-        num_configures = 6 + num_csp_ingest_configures
-        assert len(tile.Configure40GCore.mock_calls) == num_configures
-        for core in range(2):
-            assert json.loads(
-                tile.Configure40GCore.mock_calls[
-                    num_csp_ingest_configures + (3 * core)
-                ].args[0]
-            ) == {
-                "core_id": core,
-                "arp_table_entry": 0,
-                "source_port": 61648,
-                "destination_ip": (
-                    f"10.0.0.{str(154 + (2 * i) + core)}"
-                    if i != num_tiles - 1
-                    else csp_ingest_address
-                ),
-                "destination_port": csp_ingest_port,
-                "rx_port_filter": csp_ingest_port,
-                "netmask": int(ipaddress.ip_interface(sdn_first_interface).netmask),
-                "gateway_ip": int(ipaddress.ip_address(sdn_gateway)),
-            }
-            assert json.loads(
-                tile.Configure40GCore.mock_calls[
-                    num_csp_ingest_configures + 1 + (3 * core)
-                ].args[0]
-            ) == {
-                "core_id": core,
-                "arp_table_entry": 2,
-                "source_port": 61648,
-                "destination_ip": (
-                    f"10.0.0.{str(154 + (2 * i) + core)}"
-                    if i != num_tiles - 1
-                    else csp_ingest_address
-                ),
-                "destination_port": (
-                    csp_ingest_port + 2 if not last_tile else csp_ingest_port
-                ),
-                "netmask": int(ipaddress.ip_interface(sdn_first_interface).netmask),
-                "gateway_ip": int(ipaddress.ip_address(sdn_gateway)),
-            }
-            assert json.loads(
-                tile.Configure40GCore.mock_calls[
-                    num_csp_ingest_configures + 2 + (3 * core)
-                ].args[0]
-            ) == {
-                "core_id": core,
-                "arp_table_entry": 1,
-                "rx_port_filter": csp_ingest_port + 2,
-            }
+        tile.Configure40GCore.assert_not_called()
         assert len(tile.ConfigureStationBeamformer.mock_calls) == 1
         assert json.loads(tile.ConfigureStationBeamformer.mock_calls[0].args[0]) == {
             "is_first": (i == 0),
@@ -760,10 +688,10 @@ def test_Initialise(
                     "destination_ip": "10.244.170.166",
                     "destination_port": 4660,
                     "source_port": 61648,
-                    "netmask_40g": int(
+                    "netmask_40g": str(
                         ipaddress.ip_interface(sdn_first_interface).netmask
                     ),
-                    "gateway_40g": int(ipaddress.ip_address(sdn_gateway)),
+                    "gateway_40g": sdn_gateway,
                 }
             )
         )
@@ -776,10 +704,10 @@ def test_Initialise(
                     "destination_ip": "10.244.170.166",
                     "source_port": 61648,
                     "destination_port": 4660,
-                    "netmask_40g": int(
+                    "netmask_40g": str(
                         ipaddress.ip_interface(sdn_first_interface).netmask
                     ),
-                    "gateway_40g": int(ipaddress.ip_address(sdn_gateway)),
+                    "gateway_40g": sdn_gateway,
                 }
             )
         )
@@ -1022,8 +950,8 @@ def test_Standby(
                     "destination_ip": "127.0.0.1",
                     "source_port": 0xF0D0,
                     "destination_port": 4660,
-                    "netmask_40g": 4294967168,  # /25
-                    "gateway_40g": 167772414,  # 10.0.0.254
+                    "netmask_40g": "255.255.255.128",  # /25
+                    "gateway_40g": "10.0.0.254",
                 }
             ),
         ),
@@ -1038,8 +966,8 @@ def test_Standby(
                     "destination_ip": "127.0.0.1",
                     "destination_port": 4660,
                     "source_port": 0xF0D0,
-                    "netmask_40g": 4294967168,  # /25
-                    "gateway_40g": 167772414,  # 10.0.0.254
+                    "netmask_40g": "255.255.255.128",  # /25
+                    "gateway_40g": "10.0.0.254",
                 }
             ),
         ),
@@ -1282,45 +1210,7 @@ def test_SetCspIngest(
     assert station_device.cspIngestPort == 1234
     assert station_device.cspSourcePort == 2345
     for i, tile in enumerate(mock_tile_device_proxies):
-        if i != num_tiles - 1:
-            tile.Configure40GCore.assert_not_called()
-        else:
-            assert len(tile.Configure40GCore.mock_calls) == 6
-            for core in range(2):
-                assert json.loads(
-                    tile.Configure40GCore.mock_calls[(3 * core)].args[0]
-                ) == {
-                    "core_id": core,
-                    "arp_table_entry": 0,
-                    "source_ip": f"10.0.0.{str(152 + (2 * i) + core)}",
-                    "source_mac": 107752307294360 + (2 * i) + core,
-                    "source_port": 61648,
-                    "destination_ip": "123.123.234.234",
-                    "destination_port": 1234,
-                    "rx_port_filter": 1234,
-                    "netmask": int(ipaddress.ip_interface(sdn_first_interface).netmask),
-                    "gateway_ip": int(ipaddress.ip_address(sdn_gateway)),
-                }
-                assert json.loads(
-                    tile.Configure40GCore.mock_calls[1 + (3 * core)].args[0]
-                ) == {
-                    "core_id": core,
-                    "arp_table_entry": 2,
-                    "source_ip": f"10.0.0.{str(152 + (2 * i) + core)}",
-                    "source_mac": 107752307294360 + (2 * i) + core,
-                    "source_port": 61648,
-                    "destination_ip": "123.123.234.234",
-                    "destination_port": 1234,
-                    "netmask": int(ipaddress.ip_interface(sdn_first_interface).netmask),
-                    "gateway_ip": int(ipaddress.ip_address(sdn_gateway)),
-                }
-                assert json.loads(
-                    tile.Configure40GCore.mock_calls[2 + (3 * core)].args[0]
-                ) == {
-                    "core_id": core,
-                    "arp_table_entry": 1,
-                    "rx_port_filter": 1236,
-                }
+        tile.Configure40GCore.assert_not_called()
 
 
 def test_isCalibrated(station_device: SpsStation) -> None:
@@ -2074,3 +1964,36 @@ def test_healthParams(
     new_params_json = json.dumps(new_params)
     station_device.healthModelParams = new_params_json  # type: ignore[assignment]
     assert station_device.healthModelParams == new_params_json
+
+
+def test_csp_set_reset(
+    on_station_device: SpsStation,
+    mock_tile_device_proxies: list[unittest.mock.Mock],
+    change_event_callbacks: MockTangoEventCallbackGroup,
+) -> None:
+    """
+    Test the SetCspIngest and ResetCspIngest commands.
+
+    :param on_station_device: The station device to use.
+    :param mock_tile_device_proxies: mock tile proxies that have been configured with
+        the required tile behaviours.
+    :param change_event_callbacks: dictionary of Tango change event
+        callbacks with asynchrony support.
+    """
+    assert on_station_device.state() == DevState.ON
+    initial_csp_config = on_station_device.cspIngestConfig
+
+    csp_ingest_config = json.dumps(
+        {
+            "destination_ip": "123.234.345.456",
+            "source_port": 1234,
+            "destination_port": 2345,
+        }
+    )
+    rc, _ = on_station_device.SetCspIngest(csp_ingest_config)
+    assert rc == ResultCode.OK
+    assert csp_ingest_config == on_station_device.cspIngestConfig
+
+    rc, _ = on_station_device.ResetCspIngest()
+    assert rc == ResultCode.OK
+    assert initial_csp_config == on_station_device.cspIngestConfig